src/smith: size_t is not a valid OpenCL arg type

+++ src/smith/tests/data/cl/sample-1.cl
@@ -1,7 +1,7 @@
 // Sample 1: Valid OpenCL
 //
 // Initialise array with values.
-__kernel void foobar(__global int* a, const size_t nelem) {
+__kernel void foobar(__global int* a, const int nelem) {
     int id = get_global_id(0);
     if (id < nelem)
         a[id] = nelem;
